# api-document.yaml
openapi: 3.0.0
info:
  title: プロジェクトchat
  description: Lゼミのプロジェクトでチャットを作るよ！！！
  version: '1.0.0.'
servers:
  - url: 'http://localhost:8000'
    description: development application server?
  - url: 'http://localhost:3316'
    description: development db server. db software is mysql.
...
paths:
...
  /threads/{id}/archives:
    post:
      tags: 
        - 'archives'
      summary: 'スレッドのアーカイブを作成する'
      description: 'スレッドの管理者である必要がある。privateにするならpassword必須'
      parameters: 
        - $ref: '#/components/parameters/AuthByCookie'
        - name: 'id'
          in: 'path'
          required: true
          schema:
            type: 'string'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArchiveRequest'
      responses:
        '200':
          description: 'return info of archive'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetArchiveResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    # TODO: passwordの認証
    # CONT: スレッドの中身はのぞけない
    get:
      summary: 'アーカイブを取得'
      tags: 
        - 'archives'
      description: 'public: 本当に誰でも(ログインユーザじゃなくても見れる) private: パスワード知ってれば誰でも'
      parameters: 
        - $ref: '#/components/parameters/AuthByCookie'
        - name: 'id'
          in: 'path'
          required: true
          schema:
            type: 'string'
      responses:
        '200':
          description: 'return info of archive'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetArchiveResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: 'アーカイブの情報を編集する'
      tags: 
        - 'archives'
      parameters: 
        - $ref: '#/components/parameters/AuthByCookie'
        - name: 'id'
          in: 'path'
          required: true
          schema:
            type: 'string'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArchiveRequest'
      responses:
        '200':
          description: 'return info of archive'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetArchiveResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: 'アーカイブを削除する'
      tags: 
        - 'archives'
      parameters: 
        - $ref: '#/components/parameters/AuthByCookie'
        - name: 'id'
          in: 'path'
          required: true
          schema:
            type: 'string'
      responses:
        '204':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    UserResponse:
      type: "object"
      properties:
        user_id:
          type: "string"
          description: "ユーザID"
        name:
          type: "string"
          description: "ユーザネーム"
        mail:
          type: "string"
          description: "メールアドレス"
        image:
          type: "string"
          description: "アイコンのパス"
        profile:
          type: "string"
          description: "自己紹介的な"
        tags: 
          type: 'array'
          items:
            $ref: '#/components/schemas/GetTagResponse'
        evaluations:
          type: 'array'
          items:
            $ref: '#/components/schemas/GetEvaluationScoreResponse'
...
    GetThreadResponse:
      type: 'object'
      properties:
        id:
          type: 'string'
        name:
          type: 'string'
        description:
          type: 'string'
        limit_users:
          type: 'integer'
        is_public:
          type: 'integer'
        created_at:
          type: 'string'
        updated_at:
          type: 'string'
        admin:
          $ref: '#/components/schemas/UserResponse'
        tags: 
          type: 'array'
          items:
            $ref: '#/components/schemas/GetTagResponse'
...
    GetArchiveResponse:
      type: 'object'
      properties:
        id:
          type: 'string'
        path:
          type: 'string'
        is_public:
          type: 'integer'
        password:
          type: 'string'
        thread:
          $ref: '#/components/schemas/GetThreadResponse'
    UpdateArchiveRequest:
      type: 'object'
      properties:
        password:
          type: 'string'
        is_public:
          type: 'integer'
    GetTagResponse:
      type: 'object'
      properties:
        id:
          type: 'string'
        tag:
          type: 'string'
        category:
          $ref: '#/components/schemas/GetCategoryResponse'
...
    GetCategoryResponse:
      type: 'object'
      properties:
        id:
          type: 'string'
        category:
          type: 'string'
...
    GetEvaluationScoreResponse:
      type: 'object'
      properties:
        id:
          type: 'string'
        item:
          type: 'string'
        score:
          type: 'integer'
...
    ErrorResponse:
      type: 'object'
      properties:
        code:
          type: 'integer'
        message:
          type: 'string'
...
  parameters:
    AuthByCookie:
      name: "x-token"
      in: "cookie"
      description: "JWTのトークン"
      required: true
      schema:
        type: "string"
  responses:
    Success:
      description: 'return success message'
      content:
        application/json:
          schema:
            type: 'object'
            properties:
              message:
                type: 'string'
...
    BadRequest:
      description: 'bad request'
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            example: 
              code: 400
              message: 'error content'
    Unauthorized:
      description: 'unauthorized error'
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            example:
              code: 401
              message: 'error content'
    InternalServerError:
      description: 'internal server error'
      content:
        applicatoin/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
            example: 
              code: 500
              message: 'error content'